src/ai_api/api-requirements.txt
basic-pitch==0.2.6
python-dotenv==1.0.0
librosa==0.9.1
pretty_midi==0.2.10
Flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
midi2audio==0.1.1
pyFluidSynth==1.3.2